refactor:itemdisplay 파일위치 변경
