+++ frontend/src/components/ItemDisplay/ItemDisplay.styles.ts
