+++ frontend/src/components/ItemDisplay/ItemDisplay.tsx
